♻️ REFACTOR : [코드리뷰 반영] js파일 리팩토링
- 변수명 교체
- 비밀번호 min값을 상수로 선언 : 코드 가독성을 높일 수 있고, 이후 변경도 쉬울 것 같아서 변경하였습니다.
- flag 변수 타입 변경 (int->boolean)
- checkSubmitEvent 함수 로컬 변수 이용하는 방식으로 변경
